requirements.txt
Flask==2.2.3
Werkzeug<3
requests==2.28.2
beautifulsoup4==4.11.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
gunicorn==20.1.0
google-api-python-client==2.70.0
supabase-py==1.9.0
python-dotenv==1.0.0